issue with jar file fixed
the jar file created once i built the program would not run. whilst running in netbeans it would. it had something to do with the database schema not being found after running it in the command line to see the errors. could not fix however, i found a work around. i placed the jar into the general project folder rather than in the dist folder.